requirements.txt
# Core AI/LLM
ollama>=0.1.8
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
transformers>=4.40.0
torch>=2.1.0
sentence-transformers>=2.7.0

# Data Processing
pandas>=2.2.0
pyspark>=3.5.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Vector DB
chromadb>=0.5.0

# API & Backend
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
pydantic>=2.7.0

# Database
psycopg2-binary>=2.9.9
pymongo>=4.7.0
sqlalchemy>=2.0.29

# Log Processing
regex>=2024.4.16
python-dateutil>=2.9.0

# Notifications
slack-sdk>=3.27.0
aiosmtplib>=3.0.1

# Monitoring & Utils
python-dotenv>=1.0.1
loguru>=0.7.2
rich>=13.7.1
schedule>=1.2.2
watchdog>=4.0.0
